consolidate sidebar components
